requirements.txt
phidata
langchain
google-generativeai
langchain-google-genai
tavily-python
python-dotenv
langgraph 
streamlit

